update: schema and api endpoint

+++ src/main/java/com/shivam/ecommerce/product_catalog_service/model/Product.java
@@ -18,9 +18,9 @@ public class Product {
     )
     private Integer id;
     private String name;
-    private Integer price;
+    private Double price;
 
-    public Product(Integer id, String name, Integer price) {
+    public Product(Integer id, String name, Double price) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -45,11 +45,11 @@ public class Product {
         this.name = name;
     }
 
-    public Integer getPrice() {
+    public Double getPrice() {
         return price;
     }
 
-    public void setPrice(Integer price) {
+    public void setPrice(Double price) {
         this.price = price;
     }
